added workers module
